Corrected one of the bugs mentioned in Issue #7.
This commit fixes the bug with password length comparison. We were comparing 'n' number of times the received creds, where 'n' was length of received password, instead of total count credentials stored in config file. Thanks to @standa4 for reporting the bug. Using "corrected" instead of "fixed" in commit to keep the issue open until other bugs are fixed.